:hammer: code refactoring on targets
